feat: APPS-3295 APPS-3296 add collectionlayout query, split collections/[slug].vue layout in 3 components (#137)
* feat: add collectionlayout query, start layout

* fix: linter

* fix: pascal case

* fix: carryover merge change to file

* fix: component imports

* fix: dynamic component import pattern

* fix: force pascal case

* fix: imports

---------
