+++ components/ListOfItemsCollection.vue
@@ -1,0 +1,8 @@
+<script lang="ts" setup>
+</script>
+<template>
+  <main>
+    WELCOME TO LIST OF ITEMS COLLECTION PAGE
+  </main>
+</template>
+<style lang="scss" scoped></style>
