+++ components/StorytellingCollection.vue
@@ -1,0 +1,4 @@
+<script>
+// storytelling collection layout can be built here
+// then loaded into the collections/[slug].vue page using dynamic components
+</script>
